% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlFunctions.R
\name{environment_to_xml}
\alias{environment_to_xml}
\title{environment to xml}
\usage{
environment_to_xml(env)
}
\arguments{
\item{env}{An environment to convert to xml}
}
\value{
xml object for Word
}
\description{
environment to xml
}
