% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmlFunctions.R
\name{get_values}
\alias{get_values}
\title{Get the values of the document properties}
\usage{
get_values(xml)
}
\arguments{
\item{xml}{xml object}
}
\value{
Vector of values of the document properties as characters
}
\description{
Get the values of the document properties
}
