% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveProperties.R
\name{save_to_rdocx}
\alias{save_to_rdocx}
\title{Save xml object to rdocx object}
\usage{
save_to_rdocx(rdocx, xml)
}
\arguments{
\item{rdocx}{object}

\item{xml}{object}
}
\value{
nothing
}
\description{
Save xml object to rdocx object
}
