% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldFunctions.R
\name{slip_in_docprop_field}
\alias{slip_in_docprop_field}
\title{Insert property into Word document}
\usage{
slip_in_docprop_field(name, xml, doc, options = "\\\\* MERGEFORMAT", ...)
}
\arguments{
\item{name}{Property name}

\item{xml}{xml object with corresponding property definition}

\item{doc}{rdocx object}

\item{options}{field format options. (Default "\\* MERGEFORMAT" to keep formatting)}

\item{...}{Other parameters passed to officer::slip_in_seqfield()}
}
\value{
Nothing
}
\description{
Insert property into Word document
}
